First stage of refeactoring the device factors to reduce duplicate code. Better names for sensor specific interfaces. Animation support.

+++ diozero-core/src/main/java/com/diozero/api/SPIConstants.java
@@ -28,7 +28,7 @@ package com.diozero.api;
 
 
 public interface SPIConstants {
-	public static final int DEFAULT_WORD_LENGTH = 8;
+	public static final byte DEFAULT_WORD_LENGTH = 8;
 	public static final int DEFAULT_SPI_CONTROLLER = 0;
 	public static final int DEFAULT_SPI_CLOCK_FREQUENCY = 500_000;
 	public static final SpiClockMode DEFAULT_SPI_CLOCK_MODE = SpiClockMode.MODE_1;
