+++ diozero-core/src/main/java/com/diozero/sampleapps/BMP180Test.java
@@ -59,8 +59,8 @@ public class BMP180Test {
 
 			for (int i=0; i<ITERATIONS; i++) {
 				Logger.info(String.format("Temperature=%.2f, Pressure=%.2f",
-						Double.valueOf(bmp180.getTemperature()),
-						Double.valueOf(bmp180.getPressure())));
+						Float.valueOf(bmp180.getTemperature()),
+						Float.valueOf(bmp180.getPressure())));
 				SleepUtil.sleepSeconds(0.5);
 			}
 		} catch (RuntimeIOException ioe) {
